Changes made to ImageController.java
Changes Made to the following programs-
1. ImageController.java
When the user tries to edit/delete an image which is not owned by him, display error message.
The methods editImage() and deleteImage() are altered.
New method checkUser has been added.
2. Image.html
to display editError and deleteError messages on screen in case of rule violation